Image_Model/backend/requirements.txt
Flask==2.2.5
Pillow==10.0.0
numpy==1.25.0
scikit-learn==1.2.2
tensorflow==2.13.0
